Karina version/eBidProject/authentification/middlewares.py
from urllib.parse import urlparse
from channels.auth import AuthMiddlewareStack
from django.contrib.auth import get_user_model
from django.core.exceptions import PermissionDenied
from django.db import close_old_connections
from django.utils.deprecation import MiddlewareMixin
from rest_framework_simplejwt.exceptions import TokenError
