Karina version/requirements.txt
asgiref==3.8.1
channels==4.1.0
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
greenlet==3.0.3
msgpack==1.0.8
neovim==0.3.1
pillow==10.3.0
PyJWT==2.8.0
pynvim==0.5.0
sqlparse==0.4.4
typing_extensions==4.11.0
websockets==12.0
